Revert "Rydberg h factory (#652)"
This reverts commit ce1cb3515e3c508a8b21ed1306bc133fa7756012.
