+++ pyproject.toml
@@ -17,7 +17,7 @@ classifiers = [
 dependencies = [
     "juliacall>=0.9.14",
     "numpy>=1.25.2",
-    "pydantic>=1.10.13",
+    "pydantic>=1.10.12",
     "scipy>=1.9.3",
     "pandas>=2.1.0",
     "bokeh>=3.2.2",
