+++ src/bloqade/ir/__init__.py
@@ -11,7 +11,7 @@ from .control.waveform import (
     Interpolation,
     Sample,
     PythonFn,
-    to_waveform,
+    instruction,
     GaussianKernel,
     LogisticKernel,
     SigmoidKernel,
@@ -68,7 +68,7 @@ __all__ = [
     "Sample",
     "Interpolation",
     "PythonFn",
-    "to_waveform",
+    "instruction",
     "GaussianKernel",
     "LogisticKernel",
     "SigmoidKernel",
