+++ src/bloqade/ir/control/waveform.py
@@ -1,5 +1,4 @@
 from numbers import Real
-from bloqade.builder.typing import ScalarType
 from bloqade.ir.tree_print import Printer
 from bloqade.ir.scalar import (
     Scalar,
@@ -15,7 +14,6 @@ from dataclasses import InitVar
 from decimal import Decimal
 from pydantic.dataclasses import dataclass
 from beartype.typing import Any, Tuple, Union, List, Callable, Dict
-from beartype import beartype
 from enum import Enum
 
 import numpy as np
@@ -25,8 +23,7 @@ from bloqade.visualization import get_ir_figure
 from bloqade.visualization import display_ir
 
 
-@beartype
-def to_waveform(duration: ScalarType) -> Callable[[Callable], "PythonFn"]:
+def instruction(duration: Any) -> "PythonFn":
     # turn python function into a waveform instruction."""
 
     def waveform_wrapper(fn: Callable) -> "PythonFn":
@@ -488,11 +485,6 @@ class PythonFn(Instruction):
 
     def children(self):
         return {"duration": self.duration, **{p.name: p for p in self.parameters}}
-
-    def sample(
-        self, dt: ScalarType, interpolation: Union[str, "Interpolation"]
-    ) -> "Sample":
-        return Sample(self, interpolation, cast(dt))
 
 
 @dataclass
