+++ src/bloqade/ir/routine/base.py
@@ -5,8 +5,8 @@ from bloqade.ir import AtomArrangement, ParallelRegister, Sequence
 from bloqade.builder.base import Builder
 from bloqade.ir.routine.params import Params
 
-from pydantic import ConfigDict
-from pydantic.dataclasses import dataclass
+
+from dataclasses import dataclass
 from typing import TYPE_CHECKING, Union
 
 if TYPE_CHECKING:
@@ -29,10 +29,7 @@ class RoutineParse:
         return self
 
 
-__pydantic_dataclass_config__ = ConfigDict(arbitrary_types_allowed=True)
-
-
-@dataclass(frozen=True, config=__pydantic_dataclass_config__)
+@dataclass(frozen=True)
 class RoutineBase(RoutineParse):
     source: Builder
     circuit: AnalogCircuit
@@ -46,7 +43,7 @@ class RoutineBase(RoutineParse):
         return out
 
 
-@dataclass(frozen=True, config=__pydantic_dataclass_config__)
+@dataclass(frozen=True)
 class Routine(RoutineBase):
     """Result of parsing a completed Builder string."""
